Give option to keep CSV files locally and skip upload to S3
When testing pipelines on a local machine, you don't always need
to actually send the files to S3. Sometimes you just want to
verify the data was written correctly. With this change,
specifying the s3_bucket value as 'localhost' will skip the
upload to S3 and not remove the files from temp_dir.

It also skips creation of AWS session, avoiding need to have
AWS credentials for local testing.